# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCQ Online Test Platform API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/examinee/sessions/current": {
      "get": {
        "tags": [
          "examinee"
        ],
        "summary": "Get current test session",
        "description": "Return the examinee's active (in-progress) test session.\n\nIncludes progress (`answered_count`, `current_position`) and the session\nexpiry time. Use this to resume a test after a page reload.",
        "operationId": "get_current_session_api_v1_examinee_sessions_current_get",
        "responses": {
          "200": {
            "description": "Summary of the examinee's active test session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__common__SuccessResponse_ExamineeSessionResource___2"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "No active test session was found for this examinee.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/examinee/sessions/{session_id}/questions/{position}": {
      "get": {
        "tags": [
          "examinee"
        ],
        "summary": "Get question at position",
        "description": "Fetch the question at a given position (1-50) in the examinee's session.\n\nThe correct answer is never included. The response also returns the\nexaminee's previously selected option for that question, if any.",
        "operationId": "get_question_at_position_api_v1_examinee_sessions__session_id__questions__position__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "position",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The question at the requested position, with any previously selected option.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ExamineeQuestionAtPositionResource_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/examinee/sessions/{session_id}/responses/{question_id}": {
      "put": {
        "tags": [
          "examinee"
        ],
        "summary": "Save or update an answer",
        "description": "Save, change, or clear the selected option for a question.\n\nSend `selected_option: null` to clear an answer. Requires the\n`X-CSRF-Token` header. Supports optimistic concurrency via the\n`If-Unmodified-Since` header.",
        "operationId": "save_response_api_v1_examinee_sessions__session_id__responses__question_id__put",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Question Id"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional optimistic-concurrency guard (RFC 9110 HTTP-date). If the stored response was updated after this timestamp, the save is rejected with 409.",
              "title": "If-Unmodified-Since"
            },
            "description": "Optional optimistic-concurrency guard (RFC 9110 HTTP-date). If the stored response was updated after this timestamp, the save is rejected with 409."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveResponseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The saved response and the updated answered count.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_SaveResponseResult_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "409": {
            "description": "Concurrency conflict \u2014 the response was modified after the timestamp given in If-Unmodified-Since.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The resource was modified by another request.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
...
    "/api/v1/examinee/sessions/{session_id}/review": {
      "get": {
        "tags": [
          "examinee"
        ],
        "summary": "Review submitted answers",
        "description": "Return the post-submission review for a completed session.\n\nOnly available after the session has been submitted. Each item includes\nthe examinee's selected option, the correct option, and correctness.",
        "operationId": "get_review_api_v1_examinee_sessions__session_id__review_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-question review including correct answers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ReviewResult_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "404": {
            "description": "Session not found, or it has not been submitted yet.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/not-found",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "The requested resource was not found.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/v1/examinee/me/data-export": {
      "get": {
        "tags": [
          "examinee"
        ],
        "summary": "Export personal data (GDPR)",
        "description": "Download all personal data held for the authenticated examinee.\n\nReturns a JSON file as an attachment (GDPR Art. 20 data portability).\nThis endpoint bypasses the standard `SuccessResponse` envelope.",
        "operationId": "data_export_api_v1_examinee_me_data_export_get",
        "responses": {
          "200": {
            "description": "JSON file download containing all personal data held for the examinee.",
            "headers": {
              "Content-Disposition": {
                "description": "Attachment with a generated filename, e.g. attachment; filename=\"data-export.json\".",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Complete personal-data export (profile, sessions, responses)."
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SaveResponseRequest": {
        "properties": {
          "selected_option": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectedOption"
              },
              {
                "type": "null"
              }
            ],
            "description": "Selected option key (A-D), or null to clear the answer.",
            "examples": [
              "B"
            ]
          }
        },
        "type": "object",
        "title": "SaveResponseRequest",
        "description": "Answer selection for a question; null clears the answer."
      },
...
      "SelectedOption": {
        "type": "string",
        "enum": [
          "A",
          "B",
          "C",
          "D"
        ],
        "title": "SelectedOption"
      },
...
      "SuccessResponse_ExamineeQuestionAtPositionResource_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExamineeQuestionAtPositionResource",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExamineeQuestionAtPositionResource]"
      },
...
      "app__schemas__common__SuccessResponse_ExamineeSessionResource___2": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/app__schemas__examinee__ExamineeSessionResource",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExamineeSessionResource]"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Opaque server-side session token issued by the login/entry endpoints as an HttpOnly cookie. Examiner sessions last 1 hour, examinee sessions 4 hours by default."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound CSRF token required on mutating requests. Obtain it from GET /auth/me (or GET /auth/csrf before login)."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}